separated NlnN/N2 (BHC) from GMMonly clustering
