+++ include/JetSkimmer.hh
@@ -101,9 +101,6 @@ class JetSkimmer : public BaseSkimmer{
 			//_hists1D.push_back(PVdeltaT_jet_mmAvg_pred);
 			//_hists2D.push_back(e_nRhs);
 
-			nrhs_comptime->SetName("nrhs_comptime");
-			nrhs_comptime->SetTitle("nrhs_comptime");
-			graphs.push_back(nrhs_comptime);
 
 
 		};
@@ -305,23 +302,6 @@ class JetSkimmer : public BaseSkimmer{
 		TH2D* recoGenDr_recoGenDeltaTRatio = new TH2D("recoGenDr_recoGenDeltaTRatio","recoGenDr_recoGenDeltaTRatio;recoGenDr;recoGenDeltaTRatio;a.u.",25,0,1,25,0,5);
 		//13 - gen energy vs ratio of reco to gen deltaT
 		TH2D* genEnergy_recoGenDeltaTRatio = new TH2D("genEnergy_recoGenDeltaTRatio","genEnergy_recoGenDeltaTRatio;genEnergy;recoGenDeltaTRatio;a.u.",25,0,1000,25,0,5);
-
-		//comparing predicted jets + true jets
-		//TH2D* nSubClusters_nConstituents = new TH2D("nSubClusters_nConstituents", "nSubClusters_nConstituents",50,0,20,50,0,20);
-
-
-
-		//predicted jet plots
-		TH1D* nClusters = new TH1D("nClusters","nClusters",20,0,20);
-		TH1D* PVtime_pred = new TH1D("PVtime_pred","PVtime_pred",100,-10,10);	
-		//difference in tPV between two back-to-back jets
-		//previous time definitions
-		TH1D* deltaT_jet_pred = new TH1D("PVdeltaT_jet_pred","PVdeltaT_jet_pred",100,-10,10);	
-		//comp time distribution
-		TH1D* comptime = new TH1D("comptime","comptime",100,0,300);
-		//comp time as a function of number of rechits per event
-		TGraph* nrhs_comptime = new TGraph();
-
 
 		vector<timeRecoCat> trCats;
 		void MakeTimeRecoCatHists(){
@@ -737,23 +717,6 @@ class JetSkimmer : public BaseSkimmer{
 		}
 
 
-
-		void FillPredJetHists(const vector<node*>& trees){
-			int njets = 0;
-			vector<node*> cleaned_trees;
-			for(int i = 0; i < trees.size(); i++){
-				if(trees[i] == nullptr) continue;
-				//check for mirrored point - would be double counted
-				if(trees[i]->points->mean().at(1) > 2*acos(-1) || trees[i]->points->mean().at(1) < 0) continue;	
-				FillModelHists(trees[i]->model);
-				njets++;
-				cleaned_trees.push_back(trees[i]);
-			}
-			nClusters->Fill(njets);
-			FillPVHists_PredJets(cleaned_trees);
-		}
-	
-
 		//this is for one jet
 		//all hists referenced here are in hists1D
 		void FillModelHists(BasePDFMixture* model){
@@ -767,7 +730,6 @@ class JetSkimmer : public BaseSkimmer{
 			nSubClusters->Fill(nclusters);
 			model->GetNorms(norms);
 		
-			nClusters->Fill((double)nclusters);
 			//k clusters = k jets in event -> subclusters are mixture model components
 			for(int k = 0; k < nclusters; k++){
 				E_k = norms[k]/_gev;
@@ -787,33 +749,6 @@ class JetSkimmer : public BaseSkimmer{
 		
 
 
-		//find back to back jets
-		void FillPVHists_PredJets(const vector<node*>& trees){
-			int njets = (int)trees.size(); 
-			double pi = acos(-1);
-
-			double dtime, dphi, dr, phi1, t1, phi2, t2;
-			//find pairs of jets to calculate resolution	
-			//need to be back to back
-			//time of subclusters is measured as center
-			for(int i = 0; i < njets; i++){
-				if(trees[i] == nullptr) continue;
-				CalcMMAvgPhiTime(trees[i]->model, phi1, t1);
-				//PVdeltaT_jet_mmAvg_pred->Fill(t1);	
-				for(int j = i+1; j < njets; j++){
-					if(trees[i] == nullptr) continue;
-					CalcMMAvgPhiTime(trees[j]->model, phi2, t2);
-					//dphi within [pi-0.1,pi+0.1]
-					dphi = fabs(phi1 - phi2);
-					if(dphi < pi-0.1 || dphi > pi+0.1) continue;
-					//median time
-					//energy-weighted average
-					//mm average over subclusters
-					//PVdeltaT_jet_mmAvg_pred->Fill(t1 - t2);	
-				}
-			}
-
-		}
 		int FindJetPair(const vector<Jet>& injets, pair<Jet,Jet>& outjets){
 			map<double,Jet> pt_jet;
 			//map<double, int> pt_idx;
@@ -1162,21 +1097,6 @@ class JetSkimmer : public BaseSkimmer{
 			return center;
 		}
 
-		void CalcMMAvgPhiTime(BasePDFMixture* model, double& phi, double& t){
-			int kmax = model->GetNClusters();
-			phi = 0;
-			t = 0;
-			double pi, ws;
-			map<string, Matrix> params;
-			for(int k = 0; k < kmax; k++){
-				params = model->GetPriorParameters(k);
-				phi += params["pi"].at(0,0)*params["mean"].at(1,0);
-				t += params["pi"].at(0,0)*params["mean"].at(2,0);
-				ws += params["pi"].at(0,0);
-			}
-			phi /= ws;
-			t /= ws; 
-		}
 		
 		void CalcEAvgPhiTime(BasePDFMixture* model, double& phi, double& t){
 			int kmax = model->GetNClusters();
@@ -1325,14 +1245,6 @@ class JetSkimmer : public BaseSkimmer{
 			_hists2D[i]->Write();
 			}
 			//erhs_trhs->Write();		
-			for(int i = 0; i < (int)graphs.size(); i++){
-				//name = graphs[i]->GetName();
-				//TCanvas* cv = new TCanvas(name.c_str(), "");
-				//TDRGraph(graphs[i], cv, name, name, "a.u.");
-				//write cv to file			
-				//cv->Write();
-				graphs[i]->Write();
-			}
 			for(int i = 0; i < trCats.size(); i++)
 				WriteEmptyProfiles(ofile, trCats[i]);
 			WriteTimeRecoCatStack(ofile, trCats);
@@ -1343,28 +1255,7 @@ class JetSkimmer : public BaseSkimmer{
 		}
 
 
-		void SetStrategy(int i){
-			if(i == 0) _strategy = NlnN;
-			else if(i == 1) _strategy = N2;
-			else if(i == 2) _strategy = MM;
-			else return; 
-		}
-
-
 		private:
-			enum Strategy{
-				//Delauney strategy - NlnN time - for 2pi cylinder
-				NlnN = 0,
-				//traditional strategy - N^2 time
-				N2 = 1,
-				//mm only
-				MM = 2
-			};
-		
-		
-			//clustering strategy - N^2 or NlnN
-			Strategy _strategy;
-			vector<TGraph*> graphs;
 			vector<Jet> _phos; //photons for event
 
 };
